+++ src/components/layout/Layout.tsx
@@ -51,7 +51,7 @@ function Layout() {
             <BackgroundTaskRunner />
             <ToastContainer />
             <Navbar />
-            <main className="flex-1 overflow-hidden flex flex-col relative">
+            <main className="flex-1 min-h-0 overflow-hidden flex flex-col relative">
                 <Outlet />
             </main>
         </div>